[IMP] - Add confirm button in line, change field description

+++ sost_mrp_fbt/models/stock_warehouse.py
@@ -1,17 +1,24 @@
 from odoo import models, fields, api
 
 
-class ResPartner(models.Model):
+class StockWarehouse(models.Model):
     _inherit = "stock.warehouse"
 
-    show_in_mrp = fields.Boolean(string="Show In MRP", copy=False)
+    show_in_mrp = fields.Boolean(string="Show In MRP")
 
-    @api.onchange('show_in_mrp')
-    def onchange_show_in_mrp(self):
-    	if self.show_in_mrp:
-    		self.create_dynamic_fields()
-    	else:
-    		self.delete_dynamic_fields()
+
+    def write(self, vals):
+        if 'code' in vals:
+            self.delete_dynamic_fields()
+        res = super(StockWarehouse, self).write(vals)
+        if 'show_in_mrp' in vals:
+            if vals.get('show_in_mrp'):
+                self.create_dynamic_fields()
+            else:
+                self.delete_dynamic_fields()
+        if 'code' in vals and self.show_in_mrp:
+            self.create_dynamic_fields()
+        return res
 
     def create_dynamic_fields(self):
         for warehouse_id in self:
@@ -50,14 +57,16 @@ class ResPartner(models.Model):
 
     def delete_dynamic_fields(self):
         for warehouse_id in self:
-            for suffix in ['onhand_qty', 'free_qty']:
-                warehouse_code = warehouse_id.code.replace(' ', '_')
-                field_name = f"x_{warehouse_code.lower()}_{suffix}"
-                view_name = 'bom.line%s' % field_name
-                existing_field = self.env['ir.model.fields'].search([
-                    ('name', '=', field_name),
-                    ('model', '=', 'bom.wizard.line'),
-                ], limit=1)
-                view_id = self.env['ir.ui.view'].sudo().search([('name', '=', view_name)])
-                view_id.sudo().unlink()
-                existing_field.sudo().unlink()
+            if warehouse_id.code:
+                for suffix in ['onhand_qty', 'free_qty']:
+                    warehouse_code = warehouse_id.code.replace(' ', '_')
+                    field_name = f"x_{warehouse_code.lower()}_{suffix}"
+                    view_name = 'bom.line%s' % field_name
+                    existing_field = self.env['ir.model.fields'].search([
+                        ('name', '=', field_name),
+                        ('model', '=', 'bom.wizard.line'),
+                    ], limit=1)
+                    view_id = self.env['ir.ui.view'].sudo().search(
+                        [('name', '=', view_name)])
+                    view_id.sudo().unlink()
+                    existing_field.sudo().unlink()
